requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.36
alpha_vantage==2.3.1
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
rich==13.7.0
openai==1.70.0 